+++ frontend/src/app/pages/account/account-security/account-security.component.ts
@@ -36,7 +36,7 @@ export class AccountSecurityComponent implements OnInit {
   protected readonly form = this.fb.nonNullable.group(
     {
       currentPassword: ['', [Validators.required]],
-      newPassword: ['', [Validators.required, Validators.minLength(10)]],
+      newPassword: ['', [Validators.required, Validators.minLength(10), Validators.pattern(/.*\d.*/)]],
       confirmPassword: ['', [Validators.required]]
     },
     { validators: control => this.matchPasswords(control) }